Refactor way enemies appear and change paths
The enemies are now following a path very similar to the first
wave in Gyruss. I had to change the way I made enemies appear.
We now have a small "delta" value associated with each enemy.
This is a delay after which the enemy appears. Each enemy
moves _independently_ along its path segment and onto the
next and so on.